test: move some test to version specific
some of those syntax does not exist in older nodejs
