+++ babel.config.js
@@ -13,6 +13,6 @@ module.exports = {
     }],
   ],
   plugins: [
-    '@babel/plugin-proposal-class-properties',
+    '@babel/plugin-proposal-class-properties'
   ],
 };
