+++ jest.config.js
@@ -1,5 +1,9 @@
 const isCI = require('is-ci')
 
+const nodeMajorVersion = parseInt(process.version.slice(1, process.version.indexOf('.')), 10)
+const testMatch = [''].concat([14, 16].filter(v => v <= nodeMajorVersion))
+  .map(v => `**/?(*.)+(spec|test|integrate|accept|system|unit)${v}.[jt]s?(x)`)
+
 module.exports = {
   'collectCoverageFrom': [
     '<rootDir>/ts/**/*.[jt]s',
@@ -16,7 +20,7 @@ module.exports = {
     '<rootDir>/ts',
   ],
   'testEnvironment': 'node',
-  'testMatch': ['**/?(*.)+(spec|test|integrate|accept|system|unit).[jt]s?(x)'],
+  testMatch,
   'watchPlugins': [
     'jest-watch-suspend',
     'jest-watch-typeahead/filename',
