+++ ts/tersify.spec14.ts
@@ -1,0 +1,9 @@
+import { tersify } from './tersify'
+
+describe('function', () => {
+  test('nullish coalescing operator', () => {
+    function nco(s) { return s ?? 1 }
+
+    expect(tersify(nco)).toBe('fn nco(s) { return s ?? 1 }')
+  })
+})
